Download pdf files

+++ controllers/admin.js
@@ -1,16 +1,17 @@
-const Product = require('../models/product');
-const {validationResult} = require('express-validator/check')
+const Product = require("../models/product");
+const { validationResult } = require("express-validator/check");
+const fileHelper = require("../util/file");
 
-const mongoose = require('mongoose')
+const mongoose = require("mongoose");
 
 exports.getAddProduct = (req, res, next) => {
-  res.render('admin/edit-product', {
-    pageTitle: 'Add Product',
-    path: '/admin/add-product',
+  res.render("admin/edit-product", {
+    pageTitle: "Add Product",
+    path: "/admin/add-product",
     editing: false,
     hasError: false,
     errorMessage: null,
-    validationErrors: []
+    validationErrors: [],
   });
 };
 
@@ -20,59 +21,56 @@ exports.postAddProduct = (req, res, next) => {
   const price = req.body.price;
   const description = req.body.description;
   const errors = validationResult(req);
-  if (!image){
-    return res.status(422).render('admin/edit-product', {
-      pageTitle: 'Add Product',
-      path: '/admin/add-product',
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
       editing: false,
-      hasError:true,
+      hasError: true,
       product: {
         title: title,
         price: price,
-        description: description
+        description: description,
       },
-      errorMessage: 'Attached file is not an image',
-      validationErrors: errors.array()
+      errorMessage: "Attached file is not an image",
+      validationErrors: errors.array(),
     });
   }
-  
- 
-  
 
-  if ( !errors.isEmpty()) {
-    return res.status(422).render('admin/edit-product', {
-      pageTitle: 'Add Product',
-      path: '/admin/add-product',
+  if (!errors.isEmpty()) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
       editing: false,
-      hasError:true,
+      hasError: true,
       product: {
         title: title,
         price: price,
-        description: description
+        description: description,
       },
       errorMessage: errors.array()[0].msg,
-      validationErrors: []
+      validationErrors: [],
     });
   }
 
-  console.log(image)
-  const imageUrl = 'images/' + image.filename; //.replace(/\\/g, '/');
+  console.log(image);
+  const imageUrl = "images/" + image.filename; //.replace(/\\/g, '/');
   const product = new Product({
     //_id: new mongoose.Types.ObjectId('5fd26d6e700d0402d4815792'),
     title: title,
     price: price,
     description: description,
     imageUrl: imageUrl,
-    userId: req.user
+    userId: req.user,
   });
   product
     .save()
-    .then(result => {
+    .then((result) => {
       // console.log(result);
-      console.log('Created Product');
-      res.redirect('/admin/products');
+      console.log("Created Product");
+      res.redirect("/admin/products");
     })
-    .catch(err => {
+    .catch((err) => {
       // return res.status(500).render('admin/edit-product', {
       //   pageTitle: 'Add Product',
       //   path: '/admin/add-product',
@@ -88,8 +86,8 @@ exports.postAddProduct = (req, res, next) => {
       //   validationErrors: []
       // });
       //res.redirect('/500')
-      const error = new Error(err)
-      error.httpStatusCode = 500
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
@@ -97,28 +95,28 @@ exports.postAddProduct = (req, res, next) => {
 exports.getEditProduct = (req, res, next) => {
   const editMode = req.query.edit;
   if (!editMode) {
-    return res.redirect('/');
+    return res.redirect("/");
   }
   const prodId = req.params.productId;
   Product.findById(prodId)
-    .then(product => {
+    .then((product) => {
       //throw new Error('Dummy');
       if (!product) {
-        return res.redirect('/');
+        return res.redirect("/");
       }
-      res.render('admin/edit-product', {
-        pageTitle: 'Edit Product',
-        path: '/admin/edit-product',
+      res.render("admin/edit-product", {
+        pageTitle: "Edit Product",
+        path: "/admin/edit-product",
         editing: editMode,
         product: product,
-        hasError:false,
+        hasError: false,
         errorMessage: null,
-        validationErrors: []
+        validationErrors: [],
       });
     })
-    .catch(err => {
-      const error = new Error(err)
-      error.httpStatusCode = 500
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
@@ -132,79 +130,86 @@ exports.postEditProduct = (req, res, next) => {
 
   const errors = validationResult(req);
 
-  if ( !errors.isEmpty()) {
-    return res.status(422).render('admin/edit-product', {
-      pageTitle: 'Edit Product',
-      path: '/admin/edit-product',
+  if (!errors.isEmpty()) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Edit Product",
+      path: "/admin/edit-product",
       editing: true,
-      hasError:true,
+      hasError: true,
       product: {
         title: updatedTitle,
-       
+
         price: updatedPrice,
         description: updatedDesc,
-        _id:prodId
+        _id: prodId,
       },
       errorMessage: errors.array()[0].msg,
-      validationErrors: errors.array()
+      validationErrors: errors.array(),
     });
   }
 
   Product.findById(prodId)
-    .then(product => {
-      console.log(product.userId +  '_------- ' + req.user._id)
-      if ( product.userId.toString() !== req.user._id.toString()){
-        return res.redirect('/')
+    .then((product) => {
+      console.log(product.userId + "_------- " + req.user._id);
+      if (product.userId.toString() !== req.user._id.toString()) {
+        return res.redirect("/");
       }
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      if ( image ) {
-        product.imageUrl = 'images/' + image.filename; //.replace(/\\/g, '/');
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = "images/" + image.filename; //.replace(/\\/g, '/');
       }
-      
-      return product.save()
-        .then(result => {
-          console.log('UPDATED PRODUCT!');
-          res.redirect('/admin/products');
-        })
+
+      return product.save().then((result) => {
+        console.log("UPDATED PRODUCT!");
+        res.redirect("/admin/products");
+      });
     })
-    .catch(err => {
-      const error = new Error(err)
-      error.httpStatusCode = 500
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find({userId:req.user._id})
+  Product.find({ userId: req.user._id })
     // .select('title price -_id')
     // .populate('userId', 'name')
-    .then(products => {
+    .then((products) => {
       console.log(products);
-      res.render('admin/products', {
+      res.render("admin/products", {
         prods: products,
-        pageTitle: 'Admin Products',
-        path: '/admin/products'
+        pageTitle: "Admin Products",
+        path: "/admin/products",
       });
     })
-    .catch(err => {
-      const error = new Error(err)
-      error.httpStatusCode = 500
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id:prodId, userId:req.user._id})
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product){
+        return next(new Error('Product not found.'))
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
     })
-    .catch(err => {
-      const error = new Error(err)
-      error.httpStatusCode = 500
+    .then(() => {
+      console.log("DESTROYED PRODUCT");
+      res.redirect("/admin/products");
+    })
+    .catch((err) => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
